+++ alr_neo/__init__.py
@@ -1,2 +1,2 @@
 """ALRemover."""
-__version__ = "1.3.0-beta.1"
+__version__ = "1.4.0-beta.1"